Update plugins.json for the Metadata Plugin, adding npm data and adjusted description.
According to https://github.com/metalsmith/metadata they now use the @metalsmith scope. So adding it to the plugins.json file.

Noticed when hovering the entry on the website the npm command that shows still uses the old format.

Also they have added TOML as a valid file to load from so updated description as well.

Funally, updated the repo URL as that was changed slightly as well.

Please make sure I did it right. Never done this before. Based it on other entries so I think it is correct.